+++ ui/cypress/support/e2e.ts
@@ -21,7 +21,7 @@ import './commands'
 
 // Hide fetch/XHR requests from command log
 const app = window.top;
-if (!app.document.head.querySelector('[data-hide-command-log-request]')) {
+if (app && !app.document.head.querySelector('[data-hide-command-log-request]')) {
   const style = app.document.createElement('style');
   style.innerHTML = '.command-name-request, .command-name-xhr { display: none }';
   style.setAttribute('data-hide-command-log-request', '');
