+++ ui/src/app/test/page.tsx
@@ -1,0 +1,8 @@
+export default function TestPage() {
+  return (
+    <div>
+      <h1>Test Page Works!</h1>
+      <p>This is a simple test page to verify routing is working.</p>
+    </div>
+  );
+}
